% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{launchapp}
\alias{launchapp}
\title{A basic function to run a shiny app from a package}
\usage{
launchapp(app, pkg)
}
\arguments{
\item{app}{name of the app you want to run}

\item{pkg}{name of the package that app is in}
}
\description{
A basic function to run a shiny app from a package
}
\references{
this seems to be the proper way to run an app from a package
read further from Dean Atali https://stackoverflow.com/questions/37830819/developing-shiny-app-as-a-package-and-deploying-it-to-shiny-server
}
