% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushdat.R
\name{pushdat_function}
\alias{pushdat_function}
\title{A basic function that pushes a dataset to the mysql database}
\usage{
pushdat_function(
  x,
  user = Sys.getenv("db_user"),
  pass = Sys.getenv("db_pwd"),
  dbname = Sys.getenv("db_name"),
  host = Sys.getenv("db_host"),
  table
)
}
\arguments{
\item{x}{a dataframe to push}

\item{user}{account username to login to sql server}

\item{pass}{account password to login to sql server}

\item{dbname}{name of the database to login to sql server}

\item{host}{address of the sql server}

\item{table}{sql table to write to}
}
\value{
a vector with the number of rows pushed
}
\description{
A basic function that pushes a dataset to the mysql database
}
