% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{launchapp}
\alias{launchapp}
\title{A basic function to run a shiny app from a package}
\usage{
launchapp(app, pkg, launch.browser = 0, port = 3838, host = "0.0.0.0")
}
\arguments{
\item{app}{name of the app you want to run}

\item{pkg}{name of the package that app is in}

\item{launch.browser}{shiny::runApp argument wether to launch browser (defualt = 0)}

\item{port}{shiny::runApp argument to expose port (default 3838)}

\item{host}{shiny::runApp argument to define which host}
}
\description{
A basic function to run a shiny app from a package
}
\references{
this seems to be the proper way to run an app from a package
read further from Dean Atali https://stackoverflow.com/questions/37830819/developing-shiny-app-as-a-package-and-deploying-it-to-shiny-server
}
